Add eclipse support for Hedgehog Markets
